Move sorting logic out of the Continent model to a separate module.
